% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city.code.R
\docType{data}
\name{city.code}
\alias{city.code}
\title{Title of the simple function to add 1}
\format{
An object of class \code{data.frame} with 30 rows and 2 columns.
}
\usage{
city.code
}
\arguments{
\item{city.code}{wakayama city name}
}
\description{
The second paragraph is recognized as the description.
It is not recommended to use @title and @description explicitly.
}
\examples{
city.code
}
\keyword{datasets}
